client/requirements.txt
pygame>=2.1.0
nfcpy>=1.0.3
requests>=2.28.0